requirements.txt
# Core dependencies - updated for Python 3.13 compatibility
activate==0.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==2.4.1
attrs==25.3.0
bcrypt==4.0.1
blinker==1.9.0
boto3==1.39.2
botocore==1.39.2
cachetools==5.5.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.2.1
colorama==0.4.6
comm==0.2.1
debugpy==1.8.0
decorator==5.1.1
discard==0.1
discord==2.3.2
discord.py==2.5.2
Django==5.1.1
django-allauth==64.2.1
django-cors-headers==4.7.0
django-crispy-forms==2.3
django-jazzmin==3.0.1
django-resized==1.0.3
djangorestframework==3.16.0
ecdsa==0.19.1
et_xmlfile==2.0.0
executing==2.0.1
Faker==28.1.0
filelock==3.16.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.1.1
frozenlist==1.7.0
fsspec==2024.9.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grep==0.3.2
grpcio==1.74.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.25.0
idna==3.10
ipykernel==6.29.0
ipython==8.20.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.0
jupyter_core==5.7.1
langchain==0.3.25
langchain-core==0.3.65
langchain-ollama==0.3.3
langchain-text-splitters==0.3.8
langsmith==0.3.45
lxml==6.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mpmath==1.3.0
narwhals==1.45.0
nest-asyncio==1.6.0
networkx==3.3
numpy
ollama==0.5.1
opencv-contrib-python==4.8.1.78
openpyxl==3.1.5
orjson==3.10.18
packaging==23.2
pandas>=2.2.2
parso==0.8.3
pillow==10.4.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==5.9.8
pure-eval==0.2.2
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.17.2
PyJWT==2.10.1
pyparsing==3.2.3
pypdf==5.9.0
PyPDF2==3.0.1
python-dateutil==2.8.2
python-docx==1.2.0
python-dotenv==1.0.0
python-jose==3.3.0
python-pptx==1.0.2
pytz==2023.3.post1
PyYAML==6.0.2
pyzmq==25.1.2
referencing==0.36.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.26.0
rsa==4.9.1
s3transfer==0.13.0
safetensors==0.4.5
scripts==3.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.1
stack-data==0.6.3
streamlit==1.46.1
stua==0.3
sympy==1.13.3
tenacity==9.1.2
toml==0.10.2
torch>=2.5.0
tornado==6.4
tqdm==4.66.5
traitlets==5.14.1
# Updated tokenizers for Python 3.13 compatibility
tokenizers>=0.21.0
transformers>=4.45.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.2.0
urllib3==2.2.3
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
xlsxwriter==3.2.5
yarl==1.20.1
zstandard==0.23.0

# Additional dependencies
streamlit
google-generativeai
PyPDF2
python-docx
openpyxl
python-pptx
textblob
